requirements.txt
# Core
torch==2.8.0
torchvision==0.23.0
diffusers==0.35.2
transformers==4.57.1
accelerate==1.11.0
peft==0.17.1
einops==0.8.1
safetensors==0.6.2
huggingface-hub==0.35.3
tokenizers==0.22.1
pyyaml==6.0.2
tqdm==4.67.1
pillow==10.4.0
numpy==2.2.6

# UI
gradio==5.49.1

# Eval & utilities
open-clip-torch==2.32.0
torch-fidelity==0.3.0
faiss-cpu==1.12.0
pandas==2.3.2
pyarrow==21.0.0
imageio-ffmpeg==0.6.0

# Optional: speed/memory (enable only if you want it)
xformers==0.0.32.post2
